Split load & cast
