+++ compiler/annotation/expression/expression_compiler.rs
@@ -32,11 +32,12 @@ use crate::annotation::expression::{
             MathMaxDecimalDecimal, MathMaxDoubleDouble, MathMaxIntegerInteger, MathMinDecimalDecimal,
             MathMinDoubleDouble, MathMinIntegerInteger,
         },
-        list_operations,
-        load_cast::{
+        cast::{
             CastLeftDecimalToDouble, CastLeftIntegerToDecimal, CastLeftIntegerToDouble, CastRightDecimalToDouble,
-            CastRightIntegerToDecimal, CastRightIntegerToDouble, LoadConstant, LoadVariable,
+            CastRightIntegerToDecimal, CastRightIntegerToDouble,
         },
+        list_operations,
+        load::{LoadConstant, LoadVariable},
         op_codes::ExpressionOpCode,
         operators,
         unary::{
