+++ compiler/annotation/expression/instructions/mod.rs
@@ -13,8 +13,9 @@ use crate::annotation::expression::{
 };
 
 pub mod binary;
+pub mod cast;
 pub mod list_operations;
-pub mod load_cast;
+pub mod load;
 pub mod op_codes;
 pub mod operators;
 pub mod unary;
